+++ packages/sdk/examples/e2b.ts
@@ -1,9 +1,5 @@
 import { createSandbox } from "../src";
 import { e2b } from "../src/providers/e2b";
 
-const sandbox = await createSandbox({ provider: e2b() });
-try {
-  console.log((await sandbox.run("printf e2b")).stdout);
-} finally {
-  await sandbox.stop();
-}
+await using sandbox = await createSandbox({ provider: e2b() });
+console.log((await sandbox.run("printf e2b")).stdout);